Cleans up GUI code, adds HDRIs options
